+++ tests/YoTennis.Tests/Helpers/Approvals.cs
@@ -5,7 +5,7 @@ using System.IO;
 using System.Runtime.CompilerServices;
 using System.Text;
 
-namespace YoTennis.Tests.Test
+namespace YoTennis.Tests.Helpers
 {
     static class Approvals
     {
